save SVs to beacon database
